test-cases/form-test.html

<!DOCTYPE HTML>
<!-- -*- coding: utf-8 -*-  -->

<html>
  <head>
    <meta charset="UTF-8">
    <script src="resources/testharness.js"></script>
    <script src="resources/testharnessreport.js"></script>
    <script src="form-test.js"></script>
  </head>

<!-- 
Load a form into an iframe. The form auto submits, sending its form
data to the server.  The server sends a message back that contains
JSON for the request, including headers, URL, method, and the 'body'
    
testharness.js is used to check the message sent back from the server
against required values.  The results of this comparison are displayed
using the div with id="log".
-->

  <body>
    <h2>Multipart-Form-Data Form Tests</h2>
    <iframe id="form-frame" width="50%"></iframe>
    <div id="log">
      Log would go here if there were one.
    </div>

    <script>
function generateFrameContent(values) {
	var content = '<html><head><title>Form test</title><meta charset="UTF-8"></head><body>'
			+ '<form id="form" action="http://localhost:8000/common/postecho.py"'
			+ ' method="POST" enctype="multipart/form-data">';

	values.forEach(function (obj) {
		content += "<input type=text name='" + obj.fieldName
				+ "' value='" + obj.body + "'>";
		});

	return content
		+ '</form><script>document.getElementById("form").submit();<' + '/script>'
		+ '</body></html>';
	}

// sample to show testing works
// test(function() {assert_true(true)}, "assert_true with true")

// here are some real tests
var testValues = {
//	"sub-form-test": [{fieldName: "first-field", body: "Field One Text"},
//	{fieldName: "second-field", body: "Field Two Text"}]
//	,
	"sub-form-test-01": 
		[{fieldName: "first-field", body: "Field One Text"},
		{fieldName: "second-field", body: "Field Two Téxt"},
		{fieldName: "third-field", body: "Field Three Téxt"}]
};



/// how to send files from the file system (filename contains non-ascii characters)


var formFrame = document.getElementById("form-frame");

for (var testName in testValues) {
	// nested function so each test is separate
	(function testing () {
	var test = async_test("Multipart/form-data test " + testName);
	test.step(function () {
		var callback = null;
	    function receiveMessage(e) {
			var parsedParts = form_test.processResponse(JSON.parse(e.data));
			assert_object_equals(parsedParts, 
					testValues[testName],
					"multipart test " + testName);
			test.done();
			// window.removeEventListener("message", callback);
		}
		callback = test.step_func(receiveMessage);
		window.addEventListener("message", callback);
		formFrame.srcdoc = generateFrameContent(testValues[testName]);
	})
	})()
};
	</script>
</body>
</html>
